dsr_dashboard/requirements.txt
# Testing packages
pytest>=8.0.0
pytest-cov>=4.1.0  # For coverage reporting

# Core AWS packages with looser constraints
boto3>=1.34.0,<1.35.0
botocore>=1.34.0,<1.35.0
s3fs>=2024.2.0
s3transfer>=0.10.0
aiobotocore>=2.12.0
h5netcdf>=1.3.0
fsspec>=2024.2.0
urllib3>=1.25.4,<2.1

# Data processing packages
numpy>=2.2.0
pandas>=2.2.0
xarray>=2025.1.0
dask>=2025.1.0
netCDF4>=1.7.0

# Other dependencies can be more flexible
attrs>=25.1.0
cdsapi>=0.7.5
certifi>=2025.1.0
cftime==1.6.4.post1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
datapi==0.2.0
idna==3.10
jmespath==1.0.1
locket==1.0.0
lxml==5.3.1
multiurl==0.3.3
packaging==24.2
partd==1.4.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
xarray==2025.1.2
scipy>=1.11.0

# selenium testing dependencies
tornado==6.3.3
tenacity>=8.0.0
selenium>=4.0.0
Pillow>=8.0.0
pytest-selenium>=4.0.0